added become a sponsor button to sponsor section
